+++ src/view/eval.rs
@@ -48,7 +48,7 @@ impl View for EvalComponent {
                     let mut lines: Vec<Line> = Vec::new();
                     draw_properties(
                         &app.theme(),
-                        &eval_entry.response.properties,
+                        eval_entry.response.properties.defined_properties(),
                         &mut lines,
                         0,
                         &mut Vec::new(),
@@ -120,7 +120,7 @@ impl View for EvalDialog {
 
 pub fn draw_properties(
     theme: &Scheme,
-    properties: &Vec<Property>,
+    properties: Vec<&Property>,
     lines: &mut Vec<Line>,
     level: usize,
     filter_path: &mut Vec<&str>,
@@ -160,7 +160,7 @@ pub fn draw_properties(
         lines.push(Line::from(spans));
 
         if !property.children.is_empty() {
-            draw_properties(theme, &property.children, lines, level + 1, filter_path);
+            draw_properties(theme, property.children.defined_properties(), lines, level + 1, filter_path);
             lines.push(Line::from(vec![Span::raw(delimiters.1)]).style(theme.syntax_brace));
         }
     }
@@ -185,7 +185,7 @@ pub fn render_value<'a>(theme: &Scheme, property: &Property) -> Span<'a> {
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::theme::Theme;
+    use crate::{dbgp::client::Properties, theme::Theme};
     use anyhow::Result;
     use pretty_assertions::assert_eq;
 
@@ -194,7 +194,7 @@ mod test {
         let mut lines = vec![];
         draw_properties(
             &Theme::SolarizedDark.scheme(),
-            &Vec::new(),
+            Vec::new(),
             &mut lines,
             0,
             &mut Vec::new(),
@@ -209,12 +209,12 @@ mod test {
         let mut prop1 = Property::default();
         let mut prop2 = Property::default();
         prop2.name = "bar".to_string();
-        prop1.children = vec![prop2];
+        prop1.children = Properties::from_properties(vec![prop2]);
         prop1.name = "foo".to_string();
 
         draw_properties(
             &Theme::SolarizedDark.scheme(),
-            &vec![prop1],
+            vec![&prop1],
             &mut lines,
             0,
             &mut Vec::new(),
@@ -237,7 +237,7 @@ mod test {
         let prop3 = Property::default();
 
         prop2.name = "bar".to_string();
-        prop1.children = vec![prop2];
+        prop1.children = Properties::from_properties(vec![prop2]);
         prop1.name = "foo".to_string();
 
         // segments are reversed
@@ -245,7 +245,7 @@ mod test {
 
         draw_properties(
             &Theme::SolarizedDark.scheme(),
-            &vec![prop1, prop3],
+            vec![&prop1, &prop3],
             &mut lines,
             0,
             filter,